tests: move ProjectUploadComplete tests to test_project_upload.py
